Remove controller, remove wc-admin redirect, fix plan level redirect
